+++ website/src/components/basic/button/demos/icon.tsx
@@ -24,20 +24,20 @@ export default () => {
     <div className="button-demo">
       <div className="list">
         <div className="item">
-          <Button type="assist" icon={<FilterOutlined />} size="large" />
+          <Button type="text" icon={<FilterOutlined />} size="large" />
           <Button type="secondary" icon={<FilterOutlined />} size="large" />
         </div>
         <div className="item">
-          <Button type="assist" disabled icon={<FilterOutlined />} size="large" />
+          <Button type="text" disabled icon={<FilterOutlined />} size="large" />
           <Button type="secondary" disabled icon={<FilterOutlined />} size="large" />
         </div>
         <div className="item">
-          <Button type="assist" disabled icon={<FilterOutlined />} size="large" loading />
+          <Button type="text" disabled icon={<FilterOutlined />} size="large" loading />
           <Button type="secondary" disabled icon={<FilterOutlined />} size="large" loading />
         </div>
         <div className="item">
           <Button
-            type="assist"
+            type="text"
             icon={<FilterOutlined />}
             size="large"
             loading={loadings[1]}
@@ -54,38 +54,38 @@ export default () => {
       </div>
       <div className="list">
         <div className="item">
-          <Button type="assist" icon={<FilterOutlined />} />
+          <Button type="text" icon={<FilterOutlined />} />
           <Button type="secondary" icon={<FilterOutlined />} />
         </div>
         <div className="item">
-          <Button type="assist" disabled icon={<FilterOutlined />} />
+          <Button type="text" disabled icon={<FilterOutlined />} />
           <Button type="secondary" disabled icon={<FilterOutlined />} />
         </div>
         <div className="item">
-          <Button type="assist" disabled icon={<FilterOutlined />} loading />
+          <Button type="text" disabled icon={<FilterOutlined />} loading />
           <Button type="secondary" disabled icon={<FilterOutlined />} loading />
         </div>
         <div className="item">
-          <Button type="assist" icon={<FilterOutlined />} loading={loadings[3]} onClick={() => enterLoading(3)} />
+          <Button type="text" icon={<FilterOutlined />} loading={loadings[3]} onClick={() => enterLoading(3)} />
           <Button type="secondary" icon={<FilterOutlined />} loading={loadings[4]} onClick={() => enterLoading(4)} />
         </div>
       </div>
       <div className="list">
         <div className="item">
-          <Button type="assist" icon={<FilterOutlined />} size="small" />
+          <Button type="text" icon={<FilterOutlined />} size="small" />
           <Button type="secondary" icon={<FilterOutlined />} size="small" />
         </div>
         <div className="item">
-          <Button type="assist" disabled icon={<FilterOutlined />} size="small" />
+          <Button type="text" disabled icon={<FilterOutlined />} size="small" />
           <Button type="secondary" disabled icon={<FilterOutlined />} size="small" />
         </div>
         <div className="item">
-          <Button type="assist" disabled icon={<FilterOutlined />} size="small" loading />
+          <Button type="text" disabled icon={<FilterOutlined />} size="small" loading />
           <Button type="secondary" disabled icon={<FilterOutlined />} size="small" loading />
         </div>
         <div className="item">
           <Button
-            type="assist"
+            type="text"
             icon={<FilterOutlined />}
             size="small"
             loading={loadings[5]}
@@ -103,19 +103,19 @@ export default () => {
 
       <div className="list">
         <div className="item">
-          <Button type="assist" icon={<FilterOutlined />} mini />
+          <Button type="text" icon={<FilterOutlined />} mini />
           <Button type="secondary" icon={<FilterOutlined />} mini />
         </div>
         <div className="item">
-          <Button type="assist" disabled icon={<FilterOutlined />} mini />
+          <Button type="text" disabled icon={<FilterOutlined />} mini />
           <Button type="secondary" disabled icon={<FilterOutlined />} mini />
         </div>
         <div className="item">
-          <Button type="assist" disabled icon={<FilterOutlined />} mini loading />
+          <Button type="text" disabled icon={<FilterOutlined />} mini loading />
           <Button type="secondary" disabled icon={<FilterOutlined />} mini loading />
         </div>
         <div className="item">
-          <Button type="assist" icon={<FilterOutlined />} loading={loadings[5]} mini onClick={() => enterLoading(5)} />
+          <Button type="text" icon={<FilterOutlined />} loading={loadings[5]} mini onClick={() => enterLoading(5)} />
           <Button
             type="secondary"
             icon={<FilterOutlined />}