+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## [21.4.4](https://github.com/growingio/gio-design/compare/v21.4.3...v21.4.4) (2021-04-25)
+
+
+### Bug Fixes
+
+* **page:** fix page export and add component type ([#955](https://github.com/growingio/gio-design/issues/955)) ([c0d5daa](https://github.com/growingio/gio-design/commit/c0d5daaac8b5f84516ab4b924e0f64f81b07d264))
+
+
+
+## [21.4.3](https://github.com/growingio/gio-design/compare/v21.4.2...v21.4.3) (2021-04-22)
+
+
+### Features
+
+* add list pro component ([6f535c2](https://github.com/growingio/gio-design/commit/6f535c217113130a3823d04853002ec4b55d04eb))
+* **button:** update button styles  ([#946](https://github.com/growingio/gio-design/issues/946)) ([3052cfa](https://github.com/growingio/gio-design/commit/3052cfa905813348cd892b55effe3b1db9b9aa25))
+* **page:** add forbidden page ([#952](https://github.com/growingio/gio-design/issues/952)) ([6ffeb87](https://github.com/growingio/gio-design/commit/6ffeb8740d4890f8953903df88829144917e5a2e))
+
+
+### BREAKING CHANGES
+
+* **button:** button type prop update
+
+
+
 ## [21.4.2](https://github.com/growingio/gio-design/compare/v21.4.1...v21.4.2) (2021-04-16)